# Harden the MatmulTensorCore strategy