--- a/compiler/src/iree/compiler/Codegen/TransformStrategies/Common/Common.cpp
+++ b/compiler/src/iree/compiler/Codegen/TransformStrategies/Common/Common.cpp
@@ -24,6 +24,7 @@ using namespace mlir;
 
 #define DEBUG_TYPE "iree-transform-builder"
 #define DBGS() (llvm::dbgs() << "[" DEBUG_TYPE "]: ")
+#define LDBG(X) LLVM_DEBUG(DBGS() << X << "\n")
 
 // TODO: significantly better namespacing.
 using iree_compiler::IREE::transform_dialect::ApplyBufferOptimizationsOp;
@@ -107,9 +108,8 @@ void mlir::iree_compiler::createTransformRegion(
         b.create<transform::YieldOp>(loc);
       });
   (void)sequence;
-  LLVM_DEBUG(DBGS() << "transformation script:\n");
-  LLVM_DEBUG(DBGS() << "verification: " << sequence.verify().succeeded()
-                    << "\n");
+  LDBG("transformation script:\n");
+  LDBG("verification: " << sequence.verify().succeeded() << "\n");
   LLVM_DEBUG(sequence.print(DBGS()));
 }
 
@@ -296,13 +296,15 @@ Value mlir::iree_compiler::buildLowerMaskedTransfersAndCleanup(
       containingOpH, [](OpBuilder &b, Location loc) {
         b.create<transform::ApplyLowerMaskedTransfersPatternsOp>(loc);
       });
-  b.create<transform::ApplyPatternsOp>(
-      containingOpH, [](OpBuilder &b, Location loc) {
-        b.create<transform::ApplyCastAwayVectorLeadingOneDimPatternsOp>(loc);
-        b.create<transform::ApplyFoldUnitExtentDimsViaSlicesPatternsOp>(loc);
-        b.create<IREE::transform_dialect::
-                     ApplyFoldReshapeIntoTensorHalInterfacePatternsOp>(loc);
-      });
+  if (cleanup) {
+    b.create<transform::ApplyPatternsOp>(
+        containingOpH, [](OpBuilder &b, Location loc) {
+          b.create<transform::ApplyCastAwayVectorLeadingOneDimPatternsOp>(loc);
+          b.create<transform::ApplyFoldUnitExtentDimsViaSlicesPatternsOp>(loc);
+          b.create<IREE::transform_dialect::
+                       ApplyFoldReshapeIntoTensorHalInterfacePatternsOp>(loc);
+        });
+  }
   return containingOpH;
 }