@nicolasvasilache 2023-06-12
Interestingly, the UI does not surface the gains, so I am pasting manually:

Before this PR @ #14052 ([link](https://gist.github.com/iree-github-actions-bot/657d0528ab9d4e3cf45b3b9834e1e8d2)):


Benchmark Name | Average Latency (ms) | Median Latency (ms) | Latency Standard Deviation (ms)
-- | -- | -- | --
matmul\_123x2561x2561\_f32t\_f32t\_f32t\_tile\_config\_default(linalg) [cuda-sm\_80-linux\_gnu-cuda][ukernel,matmul] cuda(none)[full-inference,default-flags] with zeros @ a2-highgpu-1g[gpu] | 293.722 | 293.654 | 0.231
matmul\_2562x2561x2561\_f32t\_f32t\_f32t\_tile\_config\_default(linalg) [cuda-sm\_80-linux\_gnu-cuda][ukernel,matmul] cuda(none)[full-inference,default-flags] with zeros @ a2-highgpu-1g[gpu] | 1.443 | 1.440 | 0.010
matmul\_2562x2564x2562\_f32t\_f32t\_f32t\_tile\_config\_default(linalg) [cuda-sm\_80-linux\_gnu-cuda][ukernel,matmul] cuda(none)[full-inference,default-flags] with zeros @ a2-highgpu-1g[gpu] | 1.581 | 1.581 | 0.002
matmul\_2564x2564x2564\_f32t\_f32t\_f32t\_tile\_config\_default(linalg) [cuda-sm\_80-linux\_gnu-cuda][ukernel,matmul] cuda(none)[full-inference,default-flags] with zeros @ a2-highgpu-1g[gpu] | 1.187 | 1.187 | 0.000


With this PR ([link](https://gist.github.com/iree-github-actions-bot/bc494defbe7021734526366752cba9be)):

Benchmark Name | Average Latency (ms) | Median Latency (ms) | Latency Standard Deviation (ms)
-- | -- | -- | --
matmul\_123x2561x2561\_f32t\_f32t\_f32t\_tile\_config\_default(linalg) [cuda-sm\_80-linux\_gnu-cuda][ukernel,matmul] cuda(none)[full-inference,default-flags] with zeros @ a2-highgpu-1g[gpu] | 0.197 | 0.197 | 0.000
matmul\_2562x2561x2561\_f32t\_f32t\_f32t\_tile\_config\_default(linalg) [cuda-sm\_80-linux\_gnu-cuda][ukernel,matmul] cuda(none)[full-inference,default-flags] with zeros @ a2-highgpu-1g[gpu] | 1.350 | 1.350 | 0.000
matmul\_2562x2564x2562\_f32t\_f32t\_f32t\_tile\_config\_default(linalg) [cuda-sm\_80-linux\_gnu-cuda][ukernel,matmul] cuda(none)[full-inference,default-flags] with zeros @ a2-highgpu-1g[gpu] | 1.371 | 1.371 | 0.000
matmul\_2564x2564x2564\_f32t\_f32t\_f32t\_tile\_config\_default(linalg) [cuda-sm\_80-linux\_gnu-cuda][ukernel,matmul] cuda(none)[full-inference,default-flags] with zeros @ a2-highgpu-1g[gpu] | 1.081 | 1.077 | 0.012